requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.20.0
opencv-python>=4.5.0
Pillow>=9.0.0
matplotlib>=3.5.0
tqdm>=4.62.0
pyyaml>=6.0
scikit-image>=0.19.0
tensorboard>=2.11.0
pycocotools>=2.0.5
albumentations>=1.3.0
kornia>=0.6.8
einops>=0.6.0
timm>=0.6.0
lpips>=0.1.4
gdown>=4.6.0
pandas>=1.5.0
seaborn>=0.12.0
scipy>=1.9.0
wandb>=0.15.0 